Annotation/backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pdf2image
pillow
celery
redis
google-generativeai
python-dotenv
pydantic
